Dev (#69)
* Add BaseBacktestReporter in backtest/performance/base.py

* Refactor structure of the backtest directory

* backtest/data: replace NaN with 0 in TWSE/TPEX price cleaners

- Added `DataUtils.fill_nan(df, 0)` at the end of `clean_twse_price`
  and `clean_tpex_price`
- Ensures missing numeric values are replaced with 0 before saving CSV
- Prevents NaN from propagating into downstream backtest modules

* Add drop_table method in SQLiteUtils

* Drop price table in data.db

* refactor(loader): ensure table/db creation fallback in all loaders

- Added create_missing_tables() checks before writing to DB
- Affected: financial_statement_loader, stock_price_loader, stock_chip_loader,
            monthly_revenue_report_loader, stock_tick_loader
- Improve robustness when setup() is skipped or DB is reset

* Update end_date in update_db.py

* update update_db.py in tasks

* 1. Delete some unused importing packages
2. Remove API_LIST logical implementation in config.py, move to
   stock_tick_utils in setup_shioaji_apis method

* Delete some useless importing packages

* Delete some unused importing packages in config.py

* Update update_db.py

* update kuo_test.ipynb

* 1. Move get_last_trading_date from TimeUtils to MarketCalendar
2. Refactor get_last_trading_date

* Delete the content of api/__init__.py and refactor the way of importing apis

* Delete importing StockUtils in trader.utils/__init__.py in order to prevent it from circular importing

* Modify path of backtest results directory in config.py

* 1. Complete save_figure in StockBacktestReporter
2. Add strategy_result_dir in Backtester
3. Add python-kaleido in .yml files

* 1. Add Base class to StockBacktestReporter
2. Add setup method to StockBacktestReporter and StockBacktestAnalyzer

* feat(report): generate continuous daily equity curve with full date range

- 使用 TimeUtils.generate_date_range 建立完整日期清單
- 將 trade_records 轉換為 {date: pnl} 並累加每日損益
- 在每日迴圈中補上沒有交易的日期，沿用前一日 equity
- 確保繪製出的資產曲線連續，與回測期間每日對齊
- 保留 DataFrame 方便後續分析或平滑處理

* black to format

* Add generate_account_df in StockBacktestReporter

* refactor(report): rename and align variables for daily PnL and balance computation

- 統一變數命名為 *_value / *_list 以區分暫存變數與結果列表
- 將 daily_pnl、cumulative_pnl、balance 的命名改為語意清晰的名稱
- 使用 date 取代 d 作為迴圈變數，提升可讀性

* Fix a typo

* Rename generate_report_df -> generate_trading_report in StockBacktestReporter

* Refactor StockTradeRecord

* 1. Add save_report method in StockBacktestReporter
2. Refactor generate_trading_report

* Fix a bug about StockTradeRecord in MomentumStrategy

* fix: correct bugs in StockTradeRecord and MomentumStrategy

- Fixed incorrect attribute handling in StockTradeRecord
- Fixed bug in MomentumStrategy where open position date was misused,
  causing tuple concatenation with timedelta

* - 在 momentum_strategy.py 中添加張數計算的註解說明
- 全面檢查專案中 lot 和 share 單位使用情況
- 確認除 Ek_strategy.py 外其他模組單位使用正確
- 建議統一使用 StockUtils 轉換函數替代硬編碼"

* fix(utils): correct commission and tax rounding rules for Taiwan market

- Use int() to apply "round down" behavior for both commission and transaction tax
- Apply minimum limits: commission ≥ Commission.MinFee, tax ≥ 1
- Fully aligns with Taiwan stock market fee rules (no ceiling rounding)

* Add BACKTEST_LOGS_DIR_PATH.mkdir(parents=True, exist_ok=True) in Backtester

+++ trader/models/stock/order.py
@@ -16,7 +16,7 @@ class StockOrder:
         stock_id: str = "",
         date: datetime.datetime = None,
         price: float = 0.0,
-        volume: float = 0.0,
+        volume: int = 0,  # Unit: Lot
         position_type: PositionType = PositionType.LONG,
     ):
         # Basic Info
@@ -25,5 +25,5 @@ class StockOrder:
 
         # Order Info
         self.price: float = price  # 交易價位
-        self.volume: float = volume  # 交易股數（Unit: Shares）
+        self.volume: int = volume  # 交易張數（Unit: Lot）
         self.position_type: PositionType = position_type  # 持倉方向（Long or Short）
